llm-auditor/pyproject.toml
[project]
name = "llm-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3==1.37.37",
    "duckduckgo-search==8.0.1",
    "google-adk==0.2.0",
    "langchain-aws==0.2.21",
    "langchain-community==0.3.21",
    "litellm==1.67.0",
]
